Refactor demo CLI app and collect related commands together
